Refactor pages into index.js
