+++ daemon/web/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import App from './pages/app/App';
+import App from './pages/App';
 import store from './store';
 import './styles/index.sass';
