+++ daemon/web/package.json
@@ -12,7 +12,7 @@
     "start": "NODE_ENV=development webpack-dev-server --host=127.0.0.1",
     "build": "NODE_ENV=production webpack --config webpack.prod.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint . --ignore-pattern public/s",
+    "lint": "eslint . --ignore-pattern public/",
     "sass-lint": "sass-lint -v"
   },
   "dependencies": {
