+++ daemon/web/pages/Dashboard/index.js
